server/stablelabel-api/pyproject.toml
[project]
name = "stablelabel-api"
version = "0.1.0"
description = "StableLabel autolabelling backend — Graph API-driven sensitivity label management for MSPs"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "msal>=1.31",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "PyJWT[crypto]>=2.8",
    "python-multipart>=0.0.9",
    "arq>=0.26",
    "redis[hiredis]>=5.0,<5.3",
    "defusedxml>=0.7",
]

[project.optional-dependencies]
classifier = [
    "presidio-analyzer>=2.2",
    "spacy>=3.7",
    "pdfminer.six>=20231228",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "testcontainers[postgres,redis]>=4.0",
    "psycopg2-binary>=2.9",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["B008"]  # Depends() in function defaults is standard FastAPI pattern

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
